pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "etl-agent"
version = "0.1.0"
description = "Autonomous ETL/ELT Agent for DevOps-Driven Data Engineering"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Suresh", email = "msk88.in@gmail.com" }]

dependencies = [
  # ── LLM & Agent Framework ─────────────────────────────────────────────────
  "anthropic>=0.40.0",
  "langchain-anthropic>=0.3.0",
  "langchain-core>=0.3.0",
  "langchain-community>=0.3.0",
  "langgraph>=0.2.0",

  # ── Structured Output & Validation ────────────────────────────────────────
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",

  # ── ETL / Data ─────────────────────────────────────────────────────────────
  "pyspark>=3.5.0",
  "delta-spark>=3.2.0",
  "pandas>=2.2.0",
  "pyarrow>=17.0.0",

  # ── API Layer ──────────────────────────────────────────────────────────────
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "python-multipart>=0.0.12",
  "jinja2>=3.1.4",
  "slowapi>=0.1.9",
  "httpx>=0.27.0",

  # ── Database ───────────────────────────────────────────────────────────────
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "aiosqlite>=0.20.0",
  "asyncpg>=0.30.0",

  # ── AWS ────────────────────────────────────────────────────────────────────
  "boto3>=1.35.0",
  "s3fs>=2024.10.0",

  # ── GitHub ─────────────────────────────────────────────────────────────────
  "PyGithub>=2.5.0",

  # ── Logging & Observability ────────────────────────────────────────────────
  "structlog>=24.4.0",
  "rich>=13.9.0",

  # ── Retry & Resilience ────────────────────────────────────────────────────
  "tenacity>=9.0.0",

  # ── Caching ────────────────────────────────────────────────────────────────
  "redis>=5.2.0",

  # ── CLI ────────────────────────────────────────────────────────────────────
  "typer>=0.15.0",

  # ── Config & Serialization ─────────────────────────────────────────────────
  "pyyaml>=6.0.2",
  "python-dotenv>=1.0.1",

  # ── Packaging (for .whl generation) ───────────────────────────────────────
  "build>=1.2.0",
]

[project.scripts]
etl-agent = "etl_agent.cli:app"

[project.optional-dependencies]
dev = [
  # ── Testing ────────────────────────────────────────────────────────────────
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=6.0.0",
  "moto[s3]>=5.0.0",
  "fakeredis>=2.26.0",

  # ── Type Checking ──────────────────────────────────────────────────────────
  "mypy>=1.13.0",
  "types-PyYAML>=6.0.12",
  "types-redis>=4.6.0",
  "boto3-stubs[s3]>=1.35.0",

  # ── Linting & Formatting ───────────────────────────────────────────────────
  "ruff>=0.8.0",

  # ── Pre-commit ─────────────────────────────────────────────────────────────
  "pre-commit>=4.0.0",

  # ── Notebooks ──────────────────────────────────────────────────────────────
  "jupyter>=1.1.0",
  "nbformat>=5.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/etl_agent"]

# ── UV ─────────────────────────────────────────────────────────────────────────
[tool.uv]
dev-dependencies = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=6.0.0",
  "moto[s3]>=5.0.0",
  "fakeredis>=2.26.0",
  "mypy>=1.13.0",
  "types-PyYAML>=6.0.12",
  "types-redis>=4.6.0",
  "boto3-stubs[s3]>=1.35.0",
  "ruff>=0.8.0",
  "pre-commit>=4.0.0",
  "jupyter>=1.1.0",
  "nbformat>=5.10.0",
]

# ── Ruff ───────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
  "ARG", # flake8-unused-arguments
  "SIM", # flake8-simplify
]
ignore = ["E501", "B008", "B905"]

[tool.ruff.lint.isort]
known-first-party = ["etl_agent"]

# ── Mypy ───────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = false
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["pyspark.*", "delta.*", "langchain.*", "langgraph.*", "anthropic.*"]
ignore_missing_imports = true

# ── Pytest ─────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
  "unit: fast isolated unit tests",
  "integration: integration tests with mocked external services",
]
addopts = "--cov=src/etl_agent --cov-report=term-missing --cov-report=html --cov-fail-under=80"

# ── Coverage ───────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["src/etl_agent"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if TYPE_CHECKING:",
  "raise NotImplementedError",
]
